third_party/WebKit/LayoutTests/http/tests/inspector/fragment-expected.txt

CONSOLE MESSAGE: line 16: Done.
Tests fragment is stripped from url by resource and page agents.


Child frame url: http://127.0.0.1:8000/inspector/resources/fragment-frame.html
Child frame resource url: http://127.0.0.1:8000/inspector/resources/fragment-frame.html
Child frame request url: http://127.0.0.1:8000/inspector/resources/fragment-frame.html
